moved dashboard pop-up modal code to separate component files
